refactor: split command form modal into focused tab components
Extract the 1717-line god component into a container + 6 tab components:
- CommandFormModal.tsx (524 lines) - state management, submit logic, tab routing
- BasicTab.tsx (652 lines) - device/protocol selection, source, payload
- ParametersTab.tsx (115 lines) - command parameter definition
- ProcessingTab.tsx (192 lines) - pre-request/post-response scripting
- FramingTab.tsx (217 lines) - response framing override config
- ContextTab.tsx (110 lines) - context linking and editing
- ProtocolTab.tsx (177 lines) - Modbus/AT protocol wizard

Closes #80

via [Happy](https://happy.engineering)
